fix: rename `web_app/app` to `web_app/service`
