+++ web_app/service/edge_detector.py
