+++ web_app/service/region_detector.py
